+++ src/routers/MainRouter.ts
@@ -5,6 +5,8 @@ import {HttpResponseCodes} from "../const/HttpResponseCodes";
 import {ProxyHandler} from "../handlers/ProxyHandler";
 import {ProxyDomain} from "../domains/ProxyDomain";
 import {AuthenticationRouter} from "./AuthenticationRouter";
+import {checkJwtMiddleware} from "../middlewares/CheckJwtMiddleware";
+import {checkRoleMiddleware} from "../middlewares/CheckRoleMiddleware";
 
 export class MainRouter {
 
@@ -22,7 +24,11 @@ export class MainRouter {
         });
 
         // Admin route
-        app.use('/manager', ManagerRouter);
+        app.use(
+            '/manager',
+            [checkJwtMiddleware, checkRoleMiddleware(["ADMIN"])],
+            ManagerRouter
+        );
 
         // User account route
         app.use('/account', AuthenticationRouter);
@@ -35,10 +41,10 @@ export class MainRouter {
                     app.use(proxy.url, ProxyRouter.getRouter(proxy, logger));
                     logger.log({message: 'Route ' + proxy.url + '(' + proxy.method + ') deployed', tag: 'cluster'});
                 });
-                app.use(function (req: any, res: any, next: any) {
-                    logger.logError({message: '404 - Route ' + req.url + ' Not found.', tag: '404'});
-                    return res.status(HttpResponseCodes.NotFound).send({error: '404 - Route ' + req.url + ' Not found.'});
-                });
+            app.use(function (req: any, res: any) {
+                logger.logError({message: '404 - Route ' + req.url + ' Not found.', tag: '404'});
+                return res.status(HttpResponseCodes.NotFound).send({error: '404 - Route ' + req.url + ' Not found.'});
+            });
             }
         ).catch(err => logger.logError(err));
     }