+++ src/routers/Manager/MethodsRouter.ts
@@ -1,0 +1,87 @@
+import {Request, Response, Router} from 'express';
+import {JsonConsoleLogger} from "../../logger/JsonConsoleLogger";
+import {MethodsHandler} from "../../handlers/MethodsHandler";
+import {InputValidationException} from "../../exceptions/InputValidationException";
+import {NotFoundException} from "../../exceptions/NotFoundException";
+import {HttpResponseCodes} from "../../const/HttpResponseCodes";
+
+const router: Router = Router();
+const logger = new JsonConsoleLogger();
+const methodsHandler = new MethodsHandler();
+
+// get all dynamic methods
+router.get('/',
+    async (req: Request, res: Response) => {
+        try {
+            const response = await methodsHandler.getAll();
+            res.send(response);
+
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            logger.logError({message: e, tag: "manager"});
+            res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+        }
+    });
+
+// delete a method
+router.delete('/:api',
+    async (req: Request, res: Response) => {
+        try {
+            const api = req.params.api;
+            await methodsHandler.deleteOne(api, req.url);
+            const response = {delete: true};
+            res.send(response);
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            if (e instanceof InputValidationException) {
+                res.status(HttpResponseCodes.Conflict).send({error: e.message});
+            } else {
+                res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+            }
+            logger.logError({message: e, tag: "manager"});
+        }
+    });
+
+// create or update endpoints
+router.post('/',
+    async (req: Request, res: Response) => {
+        try {
+            const response = await methodsHandler.addOrUpdate(req.body, req.url);
+
+            res.send(response);
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            if (e instanceof InputValidationException) {
+                res.status(HttpResponseCodes.Conflict).send({error: e.message});
+            } else if (e instanceof NotFoundException) {
+                res.status(HttpResponseCodes.NotFound).send({error: e.message});
+            } else {
+                res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+            }
+            logger.logError({message: e, tag: "manager"});
+        }
+    });
+
+
+// get a method by id
+router.get('/:api',
+    async (req: Request, res: Response) => {
+        try {
+            const api = req.params.api;
+            const response = await methodsHandler.getById(api, req.url);
+
+            res.send(response);
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            if (e instanceof InputValidationException) {
+                res.status(HttpResponseCodes.Conflict).send({error: e.message});
+            } else if (e instanceof NotFoundException) {
+                res.status(HttpResponseCodes.NotFound).send({error: e.message});
+            } else {
+                res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+            }
+            logger.logError({message: e, tag: "manager"});
+        }
+    });
+
+export const MethodsRouter: Router = router;