+++ src/routers/Manager/NamespacesRouter.ts
@@ -1,0 +1,122 @@
+import {Request, Response, Router} from 'express';
+import {JsonConsoleLogger} from "../../logger/JsonConsoleLogger";
+import {NamespacesHandler} from "../../handlers/NamespacesHandler";
+import {InputValidationException} from "../../exceptions/InputValidationException";
+import {NotFoundException} from "../../exceptions/NotFoundException";
+import {HttpResponseCodes} from "../../const/HttpResponseCodes";
+
+const router: Router = Router();
+const logger = new JsonConsoleLogger();
+const namespaceHandler = new NamespacesHandler();
+
+// get all namespaces
+router.get('/',
+    async (req: Request, res: Response) => {
+        try {
+            const response = await namespaceHandler.getAll();
+            res.send(response);
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            logger.logError({message: e, tag: "manager"});
+            res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+        }
+    });
+
+// create or update namespace
+router.post('/',
+    async (req: Request, res: Response) => {
+        try {
+            const response = await namespaceHandler.addOrUpdate(req.body, req.url);
+            res.send(response);
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            if (e instanceof InputValidationException) {
+                res.status(HttpResponseCodes.Conflict).send({error: e.message});
+            } else if (e instanceof NotFoundException) {
+                res.status(HttpResponseCodes.NotFound).send({error: e.message});
+            } else {
+                res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+            }
+            logger.logError({message: e, tag: "manager"});
+        }
+    });
+
+// delete an namespace by id
+router.delete('/:api',
+    async (req: Request, res: Response) => {
+        try {
+            const api = req.params.api;
+            await namespaceHandler.deleteOne(api, req.url);
+            const response = {delete: true};
+            res.send(response);
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            logger.logError({message: e, tag: "manager"});
+            res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+        }
+    });
+
+// delete recursive namespace
+router.delete('/recursive/:api',
+    async (req: Request, res: Response) => {
+        try {
+            const api = req.params.api;
+            await namespaceHandler.deleteRecursiveOne(api, req.url);
+            const response = {delete: true};
+            res.send(response);
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            if (e instanceof InputValidationException) {
+                res.status(HttpResponseCodes.Conflict).send({error: e.message});
+            } else {
+                res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+            }
+            logger.logError({message: e, tag: "manager"});
+        }
+    });
+
+// get a namespace by id
+router.get('/:api',
+    async (req: Request, res: Response) => {
+        try {
+            const api = req.params.api;
+            const response = await namespaceHandler.getById(api, req.url);
+            res.send(response);
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            if (e instanceof InputValidationException) {
+                res.status(HttpResponseCodes.Conflict).send({error: e.message});
+            } else if (e instanceof NotFoundException) {
+                res.status(HttpResponseCodes.NotFound).send({error: e.message});
+            } else {
+                res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+            }
+            logger.logError({message: e, tag: "manager"});
+        }
+    });
+
+// build route tree
+router.get('/build/:namespace',
+    async (req: Request, res: Response) => {
+        try {
+            const api = req.params.namespace;
+            const response = await namespaceHandler.buildRoute(api, req.url);
+            res.send(response);
+
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            logger.logError({message: e, tag: "manager"});
+            res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+        }
+    });
+
+/***************************************************************************************/
+
+router.post(
+    "/swagger",
+    async (req: Request, res: Response) => {
+        await namespaceHandler.generateFromSwagger(req, res);
+    });
+
+
+export const NamespacesRouter: Router = router;