+++ src/routers/Manager/ResourcesRouter.ts
@@ -1,0 +1,119 @@
+import {Request, Response, Router} from 'express';
+import {JsonConsoleLogger} from "../../logger/JsonConsoleLogger";
+import {ResourcesHandler} from "../../handlers/ResourcesHandler";
+import {InputValidationException} from "../../exceptions/InputValidationException";
+import {NotFoundException} from "../../exceptions/NotFoundException";
+import {HttpResponseCodes} from "../../const/HttpResponseCodes";
+
+const router: Router = Router();
+const logger = new JsonConsoleLogger();
+const resourceHandler = new ResourcesHandler();
+
+// get all dynamic resources
+router.get('/',
+    async (req: Request, res: Response) => {
+        try {
+            const response = await resourceHandler.getAll();
+            res.send(response);
+
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            logger.logError({message: e, tag: "manager"});
+            res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+        }
+    });
+
+// delete a resource by id
+router.delete('/recursive/:api',
+    async (req: Request, res: Response) => {
+        try {
+            const api = req.params.api;
+            await resourceHandler.deleteOne(api, req.url);
+            const response = {delete: true};
+            res.send(response);
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            if (e instanceof InputValidationException) {
+                res.status(HttpResponseCodes.Conflict).send({error: e.message});
+            } else {
+                res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+            }
+            logger.logError({message: e, tag: "manager"});
+        }
+    });
+
+// create or update resource
+router.post('/',
+    async (req: Request, res: Response) => {
+        try {
+            const response = await resourceHandler.addOrUpdate(req.body, req.url);
+            res.send(response);
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            if (e instanceof InputValidationException) {
+                res.status(HttpResponseCodes.Conflict).send({error: e.message});
+            } else if (e instanceof NotFoundException) {
+                res.status(HttpResponseCodes.NotFound).send({error: e.message});
+            } else {
+                res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+            }
+            logger.logError({message: e, tag: "manager"});
+        }
+    });
+
+// get resource by id
+router.get('/:api',
+    async (req: Request, res: Response) => {
+        try {
+            const api = req.params.api;
+            const response = await resourceHandler.getById(api, req.url);
+            res.send(response);
+
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            logger.logError({message: e, tag: "manager"});
+            res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+        }
+    });
+
+// get an end point by Namespace id
+router.get('/namespace/:api',
+    async (req: Request, res: Response) => {
+        try {
+            const api = req.params.api;
+            const response = await resourceHandler.getTreeByNamespace(api, req.url);
+
+            res.send(response);
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            if (e instanceof InputValidationException) {
+                res.status(HttpResponseCodes.Conflict).send({error: e.message});
+            } else {
+                res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+            }
+            logger.logError({message: e, tag: "manager"});
+        }
+    });
+
+// get an end point by id
+router.get('/:api/methods',
+    async (req: Request, res: Response) => {
+        try {
+            const api = req.params.api;
+            const response = await resourceHandler.getByIdMethods(api, req.url);
+            res.send(response);
+            logger.log({managing_route: req.url, payload: req.body, response, tag: "manager"});
+        } catch (e) {
+            if (e instanceof InputValidationException) {
+                res.status(HttpResponseCodes.Conflict).send({error: e.message});
+            } else if (e instanceof NotFoundException) {
+                res.status(HttpResponseCodes.NotFound).send({error: e.message});
+            } else {
+                res.status(HttpResponseCodes.InternalServerError).send({error: e.message});
+            }
+            logger.logError({message: e, tag: "manager"});
+        }
+    });
+
+
+export const ResourcesRouter: Router = router;